Fix bad_alloc crash during global BA
In the global bundle adjuster, change the initial estimate for the
number of edges in the graph from `keyfrms.size() * lms.size()` to
`10 * lms.size()`. The previous was a large overestimation, causing
crashes due to failed memory allocation when operating on larger maps.

The new value should be a reasonable estimation (assuming that every
landmark is observed 10 times), while keeping memory consumption
acceptable.